+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 * The error with button blocking after closing the columns settings window for the `flexberry-objectlistview` component was repeated if the `folvComponentName` property was specified for the `flexberry-lookup` component.
 
+### Changed
+* Update dependency on `ember-flexberry-data` to version `2.4.0-beta.1`.
+
 ## [2.4.0-beta.2] - 2020-03-26
 ### Fixed
 * Using the `flexberry-groupedit` component with the `fixedHeader` property.