+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [2.4.0-beta.3] - 2020-04-06
 ### Added
 * The `showFiltersInModal` property in the `flexberry-objectlistview` component for displayng filters in a modal window.
 * The `errorMessageFilterActive` and `errorMessageFilters` properties for manage the filtering of errors in the `log` service.